Add 'version' field for capabilities.
If CAP_VERSION is defined 4 bits are taken from otype field to store the version (128-bit only). Otherwise the version field is 0 bits wide. This field will be used by memory versioning extension to allow tagging memory with a version, permitting prompt revocation of capabilities.

An alternative implementation would be to use bits of the address field similar to ARM MTE / TBI but we have not settled on the exact semantics of this yet.